MMB: Refactor share button
Separate logic from styling so I can re use it for other parts of the app
