+++ app/MMB/src/navigation/DetailScreen.js
@@ -5,7 +5,7 @@ import { HeaderTitle, type NavigationType } from '@kiwicom/mobile-navigation';
 import { AdaptableLayout } from '@kiwicom/mobile-shared';
 import { Translation } from '@kiwicom/mobile-localization';
 
-import HeaderShareButton from '../components/HeaderShareButton';
+import DetailShareButton from './DetailShareButton';
 import Layout from '../components/Layout';
 import MainMenuContainer from '../MainMenuContainer';
 import HelpSubmenu, { HelpSubmenuItems } from '../scenes/help';
@@ -138,7 +138,7 @@ export default class DetailsScreen extends React.Component<Props, State> {
           <Translation id="mmb.main_menu.title" />
         </HeaderTitle>
       ),
-      headerRight: <HeaderShareButton />,
+      headerRight: <DetailShareButton />,
     };
   };
